Separate CLI and lib
Closes #48
