+++ .github/workflows/ci.yml
@@ -34,6 +34,7 @@ jobs:
       - name: add library to venv
         run: |
           . venv/bin/activate
+          cd lib
           maturin develop --cargo-extra-args="--features=python"
 #      - name: Run the python tests
 #        run: |
@@ -42,6 +43,7 @@ jobs:
       - name: compile taxonomy with python bindings
         run: |
           . venv/bin/activate
+          cd lib
           maturin build --cargo-extra-args="--features=python" --release --strip --manylinux=off
 
   tests:
