+++ .gitignore
@@ -1,5 +1,4 @@
 target
-Cargo.lock
 .DS_Store
 
 *.swp
