+++ cli/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "finch"
+version = "0.4.0"
+authors = ["One Codex <vincent@onecodex.com>"]
+description = "An implementation of min-wise independent permutation locality sensitive hashing ('MinHashing') for genomic data and command-line utility for manipulation."
+keywords = ["minhash", "bioinformatics", "sketches"]
+categories = ["command-line-utilities", "science"]
+repository = "https://github.com/onecodex/finch-rs"
+license = "MIT"
+readme = "../README.md"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = "2.33.0"
+finch_lib = { version = "0.4", path = "../lib" }
+serde_json = "1"
+anyhow = "1"
+
+[dev-dependencies]
+assert_cmd = "1"
+predicates = "1"
+proptest = "0.10"
