+++ cli/tests/data/query.fa
