+++ cli/tests/test_cli.rs
