+++ lib/src/errors.rs
