+++ lib/src/lib.rs
@@ -1,12 +1,10 @@
-#[macro_use]
-extern crate serde_derive;
-
 use std::fs::File;
-use std::io::{BufReader, Read};
+use std::io::{stdin, BufReader, Read};
 use std::path::Path;
 
 use memmap::MmapOptions;
 use needletail::parse_fastx_reader;
+use needletail::parser::Format;
 use rayon::prelude::*;
 
 use crate::filtering::FilterParams;
@@ -14,7 +12,6 @@ use crate::serialization::{
     read_finch_file, read_mash_file, MultiSketch, Sketch, FINCH_BIN_EXT, FINCH_EXT, MASH_EXT,
 };
 use crate::sketch_schemes::SketchParams;
-use needletail::parser::Format;
 
 pub mod distance;
 pub mod filtering;
@@ -22,7 +19,6 @@ pub mod sketch_schemes;
 // it would be nice if there was a `pub(in main)` or something for
 // main_parsing so we don't import it for `lib` itself
 pub mod errors;
-pub mod main_parsing;
 #[cfg(feature = "python")]
 pub mod python;
 pub mod serialization;
@@ -38,16 +34,14 @@ pub fn sketch_files(
     let sketches: FinchResult<Vec<Sketch>> = filenames
         .par_iter()
         .map(|filename| {
-            // TODO: re-enable stdin if needed
             // open the file with a special case to handle stdin
-            // let sin = stdin();
-            // let reader: Box<dyn Read + Send> = if filename == &"-" {
-            //     Box::new(sin.lock())
-            // } else {
-            //     Box::new(File::open(&Path::new(filename))?)
-            // };
+            let reader: Box<dyn Read + Send> = if filename == &"-" {
+                // We're not locking it so technically not thread safe
+                Box::new(stdin())
+            } else {
+                Box::new(File::open(&Path::new(filename))?)
+            };
             // sketch!
-            let reader: Box<dyn Read + Send> = Box::new(File::open(&Path::new(filename))?);
             Ok(sketch_stream(reader, filename, sketch_params, &filters)?)
         })
         .collect();
