+++ lib/src/serialization/mash.capnp
