+++ lib/src/serialization/mash.rs
