+++ lib/src/statistics.rs
